Build: Set VERSION and DATE from git
Preparing a release involves updating the release version and date in
"configure.ac", "conman.spec", "NEWS", and "PLATFORMS".  The latter
two are primarily informational.  But the version must be updated in
"configure.ac" and "conman.spec", and the date must be updated in
"configure.ac".  Requiring this information to be manually kept in
sync across multiple files not only increases the chance of error,
but also complicates the creation of snapshot releases.

In days of yore, an attempt was made to centralize release metadata
in a single file ("META"), but that presented other problems (see
commit 547db3c).  Instead, this commit changes the build system to
generate VERSION and DATE from git.

"configure.ac" calls "build-aux/gen-version" via AC_INIT() to generate
the version string.  Similarly, it calls "build-aux/gen-date" to
generate the date string which is then substituted via AC_SUBST().
Consequently, VERSION and DATE are only updated when the "configure"
script is regenerated (e.g., by "bootstrap").

"conman.spec" is renamed to "conman.spec.in" and undergoes autoconf
variable substitution for VERSION.  But "conman.spec" needs to
exist in the release tarball in order to directly build an rpm from
it with "rpmbuild -tb".  This is remedied by copying "conman.spec"
after variable substitution to $(distdir) as part of the "dist-hook"
Makefile rule.

Snapshot releases are appended with the date and hash of the latest
commit (e.g., x.y.z.YYYYMMDD.H).  This differs from the git-describe(1)
format in that it eschews the number of commits on top of the tagged
object.  The reasons for this choice are twofold.  First, this number
is ambiguous when snapshot releases are built from multiple branches.
Consider the case of multiple commits to a given task branch that are
squashed down when merging back into master; the later commit on master
has a smaller number of commits than the tip of the task branch before
the squash.  Second, there is no mechanism to pass this number (or
the accompanying version number) via the export-subst attribute when
using git-archive(1) resulting in a different version string for the
same release depending on how it is generated -- see scenario 3 below.

The primary limitation to a YYYYMMDD format is that it presumes there
will be at most one such release per day.  If this becomes an issue,
it is straightforward to add ".HHMM" to the string.

The following scenarios are considered for both formal releases
(those with a corresponding git tag) and snapshots (those without):

  1. Git clones obtain version information from git-describe(1)
  and the commit date from git-show(1).  git-describe(1) finds the
  most recent tag reachable from the HEAD commit; the commit hash
  is appended to the descriptive string if the tag does not point to
  that commit (thereby identifying it as a snapshot release).

  2. Distribution tarballs are created via "make dist".  Consequently,
  the Makefile dist-hook is used to write ".dist-version" and
  ".dist-date" to $(distdir).  Since the dist-hook is simply echoing
  VERSION and DATE, it has already determined whether this is a formal
  or snapshot release.

  3. Git archives use the ".gitattributes" file to specify the
  "export-subst" attribute in order to perform variable expansion on
  ".gitarchive-date" and ".gitarchive-hash" -- see gitattributes(5).
  The main difficulty here is with the limited variable expansion
  available for the "--pretty=format:" option of git-log(1).
  In particular, the version information derived from git-describe(1)
  cannot be passed.  A snapshot release is presumed if the commit
  date written to ".gitarchive-date" differs from the latest date in
  the release notes file.

  4. The fallback is to check the first line of the release notes
  file ("NEWS") for the version and date of the latest release.
  Snapshot status cannot be determined in this scenario.

I've attempted to generalize "gen-version" and "gen-date" so they work
on other flavors of Unix and older versions of git.  Early attempts
used "date --file=" to read the date string before outputting it in
a different format.  However, this option is not available in the
date(1) command on FreeBSD/NetBSD/OpenBSD.  As a result, date parsing
is performed with simple string manipulation via sed(1).

The arguments for git-show(1) have evolved considerably over time.
For example:

  $ git show --no-patch --format=%cd --date=format:%F
    Fails on Debian 8.11 (git 2.1.4) with
      "fatal: unknown date format format:%F"

  $ git show --no-patch --format=%cd --date=short
  $ git show --no-patch --format=%ci
    Fails on Debian 7.11 (git 1.7.10.4) with
      "fatal: unrecognized argument: --no-patch"

  $ git show -s --format=%cd --date=short
  $ git show -s --format=%ci
    Fails on Debian 5.0.10 (git 1.5.6.5) with
      "fatal: unrecognized argument: --format=%cd" and
      "fatal: unrecognized argument: --format=%ci"

  $ git show -s --pretty=format:%ci
    Fails on Debian 4.0 (git 1.4.4.4) with
      "fatal: invalid --pretty format: format:%ci"

This last format works on current git releases as well as those going
back to at least v1.5.6.5, beyond which the fallback case will still
generate a version string but be unable to detect snapshot releases.

Reference:
- https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.70/html_node/Initializing-configure.html
- https://www.gnu.org/software/automake/manual/html_node/The-dist-Hook.html
- https://git-scm.com/book/en/v2/Customizing-Git-Git-Attributes#_export_subst

Tested:
- Arch Linux (git 2.30.0)
- CentOS Stream 8 (git 2.27.0), 8.3.2011 (git 2.27.0),
    7.9.2009 (git 1.8.3.1), 6.10 (git 1.7.1)
- Debian sid (git 2.30.0), 10.7 (git 2.20.1), 9.13 (git 2.11.0),
    8.11 (git 2.1.4), 7.11 (git 1.7.10.4), 6.0.10 (git 1.7.2.5),
    5.0.10 (git 1.5.6.5), 4.0 (git 1.4.4.4)
- Fedora 33 (git 2.29.2), 32 (git 2.26.2), 31 (git 2.25.4)
- FreeBSD 12.2 (git 2.30.0), 11.4 (git 2.30.0)
- NetBSD 9.1 (git 2.29.2), 9.0 (git 2.29.2), 8.1 (git 2.28.0)
- OpenBSD 6.8 (git 2.28.0), 6.7 (git 2.26.2), 6.6 (git 2.24.2)
- openSUSE 15.2 (git 2.26.2), 15.1 (git 2.26.2)
- Ubuntu 20.10 (git 2.27.0), 20.04.1 LTS (git 2.25.1),
    18.04.5 LTS (git 2.17.1), 16.04.7 LTS (git 2.7.4),
    14.04.6 LTS (git 1.9.1), 12.04.5 LTS (git 1.7.9.5)