Build: Add runstatedir
Modern Linux systems typically store runtime data in "/run" instead
of the traditional "/var/run".  This affects the pidfile path.

Switch to "${runstatedir}" for storing modifiable per-process data
(e.g., the pidfile).  This variable was added in autoconf-2.70
and defaults to "${localstatedir}/run".  It can be set with the
"--runstatedir" configure option:

  ./configure --runstatedir=/run

If ${runstatedir} is undefined (i.e., if the configure script was
generated by an earlier version of autoconf), set it to the traditional
default value.  It can also be set by appending the variable assignment
to the configure line if "--runstatedir" is not recognized:

  ./configure runstatedir=/run

Update the spec file to append "runstatedir=%{_rundir}" to the
%configure line since the "--runstatedir" option is not supported on
older RedHat-based distros.  The "--runstatedir" option can be used
if the configure script was built using a newer version of autoconf
(as in the case with the release tarball), but the "runstatedir"
variable assignment will work everywhere.

Reference:
- https://lwn.net/Articles/436012/
- https://www.gnu.org/prep/standards/html_node/Directory-Variables.html
- https://lists.gnu.org/archive/html/bug-autoconf/2013-09/msg00002.html
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=759647
- https://docs.fedoraproject.org/en-US/packaging-guidelines/RPMMacros/

Tested:
Tested by running "./bootstrap" on the test system to regenerate
the configure script, running "./configure" with no args to see the
default ${runstatedir}, running "./configure --localstatedir=/var"
to verify the default ${runstatedir} was updated accordingly, running
"./configure --runstatedir=/run" to see if the option was supported,
and running "./configure runstatedir=/run" to verify the variable
could still be set if the option was not supported by the local
version of autoconf.

Success with setting the option ("--runstatedir=/run"):
- AlmaLinux 9.0
- Arch Linux
- CentOS Linux Stream 9
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.5, 10.13, 9.13
- Fedora 36, 35, 34
- FreeBSD 13.1, 13.0, 12.3
- NetBSD 9.3, 9.2, 9.1, 9.0
- OpenBSD 7.1, 7.0, 6.9
- Ubuntu 22.04.1, 20.04.5, 18.04.6, 16.04.7

Success with setting the variable ("runstatedir=/run"):
- AlmaLinux 9.0, 8.6
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.5, 10.13, 9.13, 8.11, 7.11, 6.0.10
- Fedora 36, 35, 34
- FreeBSD 13.1, 13.0, 12.3
- NetBSD 9.3, 9.2, 9.1, 9.0
- OpenBSD 7.1, 7.0, 6.9
- openSUSE 15.4, 15.3, 15.2, 15.1
- Ubuntu 22.04.1, 20.04.5, 18.04.6, 16.04.7, 14.04.6, 12.04.5

Failed ./bootstrap since the local automake version was too old:
- Debian GNU/Linux 5.0.10, 4.0