Spec: Update BuildRequires for systemd-rpm-macros
Update the BuildRequires needed for packaging systemd unit files.
This is required for the %{_unitdir} macro to exist.

"systemd-rpm-macros" is used on EPEL 8 and higher, but "systemd" is
used on EPEL 7.  Instead of an %if/%else/%endif, use a conditional
expansion based on the %{el7} variable that will only be true
(i.e., defined to 1) on RHEL 7 builds.

Reference:
- https://docs.fedoraproject.org/en-US/packaging-guidelines/Systemd/#packaging_filesystem
- https://fedoraproject.org/wiki/Packaging:Systemd#Filesystem_locations
- https://www.mail-archive.com/epel-devel@lists.fedoraproject.org/msg09325.html
- https://fedoraproject.org/wiki/Packaging:DistTag
- https://backreference.org/2011/09/17/some-tips-on-rpm-conditional-macros/index.html

Tested:
- AlmaLinux 9.0, 8.6
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009
- Fedora 36, 35, 34