Build: Add configure opts for systemd/sysvinit dirs
Add the following configure options to specify the location of
installation directories for systemd and sysvinit:

  --with-systemdunitdir[=DIR] / --without-systemdunitdir
    systemd service file installation directory

  --with-sysvinitddir[=DIR] / --without-sysvinitddir
    sysvinit script installation directory

  --with-sysconfigdir[=DIR] / --without-sysconfigdir
    sysvinit config file installation directory

These options allow the installation to be more fully configured
without having to move/delete installed files.

Update the Makefile rules for variable substitution and installation
of the init files to use the new variables.

Switch from using the "install-data-local" target to
"install-data-hook" which is always executed after the main rule's
work is done; with "-local" targets, there is no particular guarantee
of execution order.

Update the spec file (which targets RHEL7+) to use "--systemdunitdir".
In %install, remove the "rm -f" of the sysvinit files since they are
no longer installed on RHEL7+ systems which use systemd.

Reference:
- https://autotools.io/autoconf/arguments.html
- https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.71/html_node/External-Software.html
- https://www.gnu.org/software/automake/manual/html_node/Extending.html

Tested:
- AlmaLinux 9.0, 8.6
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.5, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 36, 35, 34
- FreeBSD 13.1, 13.0, 12.3
- NetBSD 9.3, 9.2, 9.1, 9.0
- OpenBSD 7.1, 7.0, 6.9
- openSUSE 15.4, 15.3, 15.2, 15.1
- Ubuntu 22.04.1, 20.04.5, 18.04.6, 16.04.7, 14.04.6, 12.04.5