conmen: Add -T opt to specify terminal emulator
Add the "-T <program>" command-line option to specify the terminal
emulator to be spawned for each console in the broadcast session.
xterm is still used by default.

Note that while an alternate terminal emulator can be specified, the
code assumes it will support options for "-geometry <geometry>" (for
the X window geometry such as "80x24" which can be specified with the
"-G" option), "-T <string>" (for the window title), and "-e <program>"
(for the program to be run in the new window).  In my limited testing,
these options are supported by xterm, lxterm, mate-terminal.wrapper
(but not mate-terminal), and urxvt.