Build: Append trailing comment line to makefile vars
Append a line with a tab-indented comment to Makefile variables.
For variables spanning multiple lines, this allows all lines to
end with a trailing backslash without having to treat the last line
differently in order to avoid the error "blank line following trailing
backslash".  This trailing comment makes it less error-prone to sort
or re-arrange lines within a variable definition.