Sharness: Run conmand on ephemeral port
If conmand is already running on the default TCP port, sharness tests
attempting to start conmand will fail with:

> ERROR: Unable to bind to port 7890: Address already in use

This can also occur if multiple sharness tests are run concurrently
(e.g., via "make -j").

Allow the conmand port (specified by config file or command line) to
be 0, in which case an ephemeral port will be bound to the listening
socket.  Write a log message indicating the port number in use.

For sharness tests, start conmand on an ephereral port, extract the
listening socket port number from the logfile, and define CONMAND_PORT
for use by the conman client to connect to conmand.

Tested:
- AlmaLinux 9.1, 8.7
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.5, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 37, 36, 35
- FreeBSD 13.1, 13.0, 12.3
- NetBSD 9.3, 9.2, 9.1, 9.0
- OpenBSD 7.2, 7.1, 7.0
- openSUSE 15.4, 15.3, 15.2, 15.1
- Ubuntu 22.10, 22.04.1, 20.04.5, 18.04.6, 16.04.7, 14.04.6, 12.04.5