Sharness: Replace seq with test_seq
In "0001-basic.t", the "stop conmand" test failed on OpenBSD 7.0 with:

> seq: not found

OpenBSD 7.0 (and prior) is missing seq(1) for printing a sequence of
numbers.  seq(1) is used in conmand_stop() to loop up to a specified
number of times polling the process listing waiting for the given
pid to exit.

Sharness provides test_seq() for portability since seq is not in
POSIX and GNU seq might not be available everywhere.

Replace "seq" with "test_seq".

Reference:
- https://www.openbsd.org/plus71.html

Tested:
- AlmaLinux 9.1, 8.7
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.5, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 37, 36, 35
- FreeBSD 13.1, 13.0, 12.3
- NetBSD 9.3, 9.2, 9.1, 9.0
- OpenBSD 7.2, 7.1, 7.0
- openSUSE 15.4, 15.3, 15.2, 15.1
- Ubuntu 22.10, 22.04.1, 20.04.5, 18.04.6, 16.04.7, 14.04.6, 12.04.5