Sharness: Remove hard-coded values for console count
Remove the hard-coded values for the number of consoles defined in
a test configuration.

In "sharness.d/03-conmand.sh", define CONMAND_CONSOLE_COUNT to
contain the computed value for the number of consoles in its config.
In "0001-basic.t", replace the hard-coded value with this variable.

In "1000-chaos-rpm.t", replace its hard-coded value with a shell cmd
to dynamically compute the value for its config.

Tested:
- AlmaLinux 9.1, 8.7
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.6, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 37, 36, 35
- FreeBSD 13.1, 12.4, 12.3
- NetBSD 9.3, 9.2, 9.1
- OpenBSD 7.2, 7.1, 7.0
- openSUSE 15.4, 15.3, 15.2, 15.1
- Ubuntu 22.10, 22.04.1, 20.04.5, 18.04.6, 16.04.7, 14.04.6, 12.04.5